+++ examples/src/FPS_TEST/main.cpp
@@ -1,24 +1,31 @@
 
 // Example file name : main.cpp
 // Description:
-// Test file for ER_OLEDM1_CH1115 library, showing use of mulitple buffers.  
-// In this case: two, divided horizontally.
-// The user can also divide vertically. and create as many buffers as they want.
+// Test file for ER_OLEDM1_CH1115 library, FPS TEST
 // URL: https://github.com/gavinlyonsrepo/ER_OLEDM1_CH1115_RPI
+//
+// Test results values SPICLK_FREQ ::
+// 256 = BCM2835_SPI_CLOCK_DIVIDER_256 =  120 FPS , ~1.6 MHz RPI3
+// 64  = default(BCM2835_SPI_CLOCK_DIVIDER_64) = 270 FPS ,~6 MHZ RPI3
+// 8 = BCM2835_SPI_CLOCK_DIVIDER_8 = 470 FPS , ~50 MHZ RPI3
+
 // *****************************
 
 #include <bcm2835.h>
-#include "ER_OLEDM1_CH1115.h"
 #include <time.h>
 #include <stdio.h>
+#include "ER_OLEDM1_CH1115.hpp"
 
-#define OLEDcontrast 0x80 //Constrast 00 to FF , user adjust
-#define myOLEDwidth  128
-#define myOLEDheight 64
 
-// GPIO
-#define RES 25 // GPIO pin number pick any you want
-#define DC 24 // GPIO pin number pick any you want
+// ==== Globals ====
+const uint8_t RES = 25; // GPIO pin number pick any you want
+const uint8_t DC = 24; // GPIO pin number pick any you want 
+const uint8_t myOLEDwidth  = 128;
+const uint8_t myOLEDheight = 64;
+
+const uint32_t SPICLK_FREQ = 64; // Spi clock divider see bcm2835SPIClockDivider enum bcm2835
+const uint8_t SPI_CE_PIN = 0; // which HW SPI chip enable pin to use,  0 or 1
+const uint8_t OLEDcontrast = 0x80; //Constrast 00 to FF , 0x80 is default.
 
 ERMCH1115 myOLED(myOLEDwidth ,myOLEDheight , RES, DC); // instantiate  an object
 
@@ -30,8 +37,7 @@ uint64_t  previousCounter =0;
 // =============== Function prototype ================
 void setup(void);
 void myTest(void);
-void display_Left(MultiBuffer* , long , int );
-void display_Right(MultiBuffer* );
+void display(long , int );
 static uint64_t counter( void );
 
 // ======================= Main ===================
@@ -39,6 +45,7 @@ int main(int argc, char **argv)
 {
 	if(!bcm2835_init())
 	{
+		printf("OLED :: ERROR failed to init bcm2835 library.\r\n");
 		return -1;
 	}
 
@@ -57,33 +64,23 @@ void setup()
 {
 	bcm2835_delay(50);
 	printf("OLED Begin\r\n");
-	myOLED.OLEDbegin(OLEDcontrast); // initialize the OLED
+	myOLED.OLEDbegin(OLEDcontrast, SPICLK_FREQ, SPI_CE_PIN); // initialize the OLED
 	myOLED.OLEDFillScreen(0x03);
 	bcm2835_delay(2400);
 }
 
 void myTest() {
-	
-
 	myOLED.setTextColor(FOREGROUND);
 	myOLED.setTextSize(1);
-	uint8_t  screenBuffer[(myOLEDwidth * (myOLEDheight / 8)) / 2]; //(128 * 8)/2 = 512 bytes
-
-	MultiBuffer left_side;
-	// Intialise  struct with buffer details (&struct,  buffer, w, h, x-offset,y-offset)
-	myOLED.OLEDinitBufferStruct(&left_side, screenBuffer, myOLEDwidth/2, myOLEDheight, 0, 0);
-	
-	MultiBuffer right_side;
-	// Intialise struct with buffer details (&struct,  buffer, w, h, x-offset,y-offset)
-	myOLED.OLEDinitBufferStruct(&right_side, screenBuffer, myOLEDwidth/2, myOLEDheight, myOLEDwidth/2, 0);
+	// Buffer setup, Define a buffer to cover whole screen
+	uint8_t screenBuffer[(myOLEDwidth * (myOLEDheight / 8))+1]; // 1024 bytes = 128 * 64/8
+	myOLED.OLEDbuffer = (uint8_t*) &screenBuffer;  // Assign the pointer to the buffer
+	myOLED.OLEDclearBuffer(); // Clear buffer
 
 	while (count < 10000)
 	{
 		static long framerate = 0;
-		display_Left(&left_side, framerate, count);
-
-		display_Right(&right_side);
-
+		display(framerate, count);
 		framerate++;
 		count++;
 		bcm2835_delay(1);
@@ -93,21 +90,17 @@ void myTest() {
 
 
 // Function to display left hand side buffer
-void display_Left(MultiBuffer* targetbuffer, long currentFramerate, int count)
+void display(long currentFramerate, int count)
 {
-	myOLED.ActiveBuffer = targetbuffer; // set target buffer object
 	myOLED.OLEDclearBuffer();
-	myOLED.setCursor(0, 0);
-	myOLED.print("LHS Buffer");
 
 	myOLED.setCursor(0, 10);
-	myOLED.print("512 bytes");
-
+	myOLED.print("1024 bytes");
 	myOLED.setCursor(0, 20);
 	myOLED.print("G Lyons");
-
 	myOLED.setCursor(0, 30);
 	myOLED.print(count);
+	
 	// Values to count frame rate per second
 	static long lastFramerate = 0;
 	static uint16_t fps;
@@ -125,23 +118,14 @@ void display_Left(MultiBuffer* targetbuffer, long currentFramerate, int count)
 	myOLED.print(fps);
 	myOLED.print(" fps");
 	myOLED.setCursor(0, 50);
-	myOLED.print("V 1.3");
-	myOLED.drawFastVLine(92, 0, 63, FOREGROUND);
-	myOLED.OLEDupdate();
-}
-
-// Function to display right hand side buffer
-void display_Right(MultiBuffer* targetbuffer)
-{
-	myOLED.ActiveBuffer = targetbuffer; // set target buffer object
-	myOLED.OLEDclearBuffer();
-	myOLED.setCursor(0, 0);
-	myOLED.print("RHS buffer");
-
-	myOLED.fillRect(0, 10, 20, 20, colour);
-	myOLED.fillCircle(40, 20, 10, !colour);
-	myOLED.drawRoundRect(10, 40, 40, 20, 10, FOREGROUND);
-	myOLED.drawPixel(60, 60, FOREGROUND);
+	myOLED.print("V 1.3.2");
+	
+	myOLED.drawFastVLine(64, 0, 63, FOREGROUND);
+	myOLED.fillRect(70, 10, 20, 20, colour);
+	myOLED.fillCircle(110, 20, 10, !colour);
+	myOLED.drawRoundRect(80, 40, 40, 20, 10, FOREGROUND);
+	myOLED.drawPixel(65, 60, FOREGROUND);
+	
 	myOLED.OLEDupdate();
 }