+++ examples/src/SOFTWARE_SPI/main.cpp
@@ -5,26 +5,29 @@
 // URL: https://github.com/gavinlyonsrepo/ER_OLEDM1_CH1115_RPI
 // *****************************
 // NOTES :
-// (1) Speed test results measured frame rate 93 fps 1:49 to  10000
+// (1) Speed test results measured frame rate 93 fps, 1:49s to count 10000
 // (2) This is SOFTWARE SPI
 // ******************************
 // 
 
 #include <bcm2835.h>
-#include "ER_OLEDM1_CH1115.h"
 #include <time.h>
 #include <stdio.h>
+#include "ER_OLEDM1_CH1115.hpp"
 
-#define OLEDcontrast 0x80 // Constrast 00 to FF 
-#define myOLEDwidth  128
-#define myOLEDheight 64
+
+// ==== Globals ====
+
+const uint8_t myOLEDwidth  = 128;
+const uint8_t myOLEDheight = 64;
+const uint8_t OLEDcontrast = 0x80; //Constrast 00 to FF , 0x80 is default.
 
 // GPIO
-#define RES 25 // GPIO pin number pick any you want
-#define DC 24 // GPIO pin number pick any you want
-#define DIN 5 // GPIO pin number pick any you want
-#define SCLK 6 // GPIO pin number pick any you want
-#define CS 8 // GPIO pin number pick any you want
+const uint8_t RES = 25; // GPIO pin number pick any you want
+const uint8_t DC = 24; // GPIO pin number pick any you want
+const uint8_t DIN = 5; // GPIO pin number pick any you want
+const uint8_t SCLK = 6; // GPIO pin number pick any you want
+const uint8_t CS = 8; // GPIO pin number pick any you want
 
 
 // software SPI
@@ -38,8 +41,7 @@ uint64_t  previousCounter =0;
 // =============== Function prototype ================
 void setup(void);
 void myLoop(void);
-void display_Left(MultiBuffer* , long , int );
-void display_Right(MultiBuffer* );
+void display(long, int );
 static uint64_t counter( void );
 
 // ======================= Main ===================
@@ -47,6 +49,7 @@ int main(int argc, char **argv)
 {
 	if(!bcm2835_init())
 	{
+		printf("OLED :: ERROR failed to init bcm2835 library.\r\n");
 		return -1;
 	}
 
@@ -76,23 +79,16 @@ void myLoop() {
 
 	myOLED.setTextColor(FOREGROUND);
 	myOLED.setTextSize(1);
-	uint8_t  screenBuffer[(myOLEDwidth * (myOLEDheight / 8)) / 2]; //(128 * 8)/2 = 512 bytes
-
-	MultiBuffer left_side;
-	// Intialise  struct with buffer details (&struct,  buffer, w, h, x-offset,y-offset)
-	myOLED.OLEDinitBufferStruct(&left_side, screenBuffer, myOLEDwidth/2, myOLEDheight, 0, 0);
 	
-	MultiBuffer right_side;
-	// Intialise struct with buffer details (&struct,  buffer, w, h, x-offset,y-offset)
-	myOLED.OLEDinitBufferStruct(&right_side, screenBuffer, myOLEDwidth/2, myOLEDheight, myOLEDwidth/2, 0);
+	// Buffer setup, Define a buffer to cover whole screen
+	uint8_t screenBuffer[(myOLEDwidth * (myOLEDheight / 8))+1]; // 1024 bytes = 128 * 64/8
+	myOLED.OLEDbuffer = (uint8_t*) &screenBuffer;  // Assign the pointer to the buffer
+	myOLED.OLEDclearBuffer(); // Clear buffer
 
 	while (count < 10000)
 	{
 		static long framerate = 0;
-		display_Left(&left_side, framerate, count);
-
-		display_Right(&right_side);
-
+		display(framerate, count);
 		framerate++;
 		count++;
 		bcm2835_delay(1);
@@ -102,19 +98,14 @@ void myLoop() {
 
 
 // Function to display left hand side buffer
-void display_Left(MultiBuffer* targetbuffer, long currentFramerate, int count)
+void display(long currentFramerate, int count)
 {
-	myOLED.ActiveBuffer = targetbuffer; // set target buffer object
 	myOLED.OLEDclearBuffer();
-	myOLED.setCursor(0, 0);
-	myOLED.print("LHS Buffer");
-
+	
 	myOLED.setCursor(0, 10);
-	myOLED.print("512 bytes");
-
+	myOLED.print("1024 bytes");
 	myOLED.setCursor(0, 20);
 	myOLED.print("G Lyons");
-
 	myOLED.setCursor(0, 30);
 	myOLED.print(count);
 	// Values to count frame rate per second
@@ -134,25 +125,16 @@ void display_Left(MultiBuffer* targetbuffer, long currentFramerate, int count)
 	myOLED.print(fps);
 	myOLED.print(" fps");
 	myOLED.setCursor(0, 50);
-	myOLED.print("V 1.3");
-	myOLED.drawFastVLine(92, 0, 63, FOREGROUND);
+	myOLED.print("V 1.3.2");
+	
+	myOLED.drawFastVLine(64, 0, 63, FOREGROUND);
+	myOLED.fillRect(70, 10, 20, 20, colour);
+	myOLED.fillCircle(110, 20, 10, !colour);
+	myOLED.drawRoundRect(80, 40, 40, 20, 10, FOREGROUND);
+	
 	myOLED.OLEDupdate();
 }
 
-// Function to display right hand side buffer
-void display_Right(MultiBuffer* targetbuffer)
-{
-	myOLED.ActiveBuffer = targetbuffer; // set target buffer object
-	myOLED.OLEDclearBuffer();
-	myOLED.setCursor(0, 0);
-	myOLED.print("RHS buffer");
-
-	myOLED.fillRect(0, 10, 20, 20, colour);
-	myOLED.fillCircle(40, 20, 10, !colour);
-	myOLED.drawRoundRect(10, 40, 40, 20, 10, FOREGROUND);
-
-	myOLED.OLEDupdate();
-}
 
 //This returns nano-seconds as a 64-bit unsigned number, monotonically increasing, 
 //probably since system boot.