+++ include/ER_OLEDM1_CH1115_graphics.hpp
@@ -1,16 +1,15 @@
 /*
 * Project Name: ER_OLEDM1_CH1115
-* File: ER_OLEDM1_CH1115_graphics.h
+* File: ER_OLEDM1_CH1115_graphics.hpp
 * Description: ER_OLEDM1 OLED driven by CH1115 controller header file 
 * for the graphics functions.
-* Author: Gavin Lyons.
 * URL: https://github.com/gavinlyonsrepo/ER_OLEDM1_CH1115_RPI
 */
 
 #ifndef _ER_OLEDM1_CH1115_GRAPHICS_H
 #define _ER_OLEDM1_CH1115_GRAPHICS_H
 
-#include "ER_OLEDM1_CH1115_Print.h"
+#include "ER_OLEDM1_CH1115_Print.hpp"
 #include <stdio.h>
 #include <stdint.h>
 
@@ -18,30 +17,39 @@
 
 typedef enum 
 {
-    CH1115Font_Default = 1,
-    CH1115Font_Thick = 2,
-    CH1115Font_Seven_Seg = 3,
-    CH1115Font_Wide = 4,
-    CH1115Font_Bignum = 5,
-    CH1115Font_Mednum = 6,
-}OLED_FONT_TYPE_e;
+	OLEDFontType_Default = 1,
+	OLEDFontType_Thick = 2, // NO LOWERCASE
+	OLEDFontType_SevenSeg = 3,
+	OLEDFontType_Wide = 4, // NO LOWERCASE
+	OLEDFontType_Tiny = 5,
+	OLEDFontType_Homespun = 6,
+	OLEDFontType_Bignum = 7, // NUMBERS + : . ,one size
+	OLEDFontType_Mednum = 8   // NUMBERS + : . ,one size
+}OLEDFontType_e;
 
 typedef enum 
 {
-	FONT_W_5 = 5, FONT_W_7 = 7, FONT_W_4 = 4, FONT_W_8 = 8,FONT_W_16= 16
-}OLED_Font_width_e; // width of the font in bytes cols.
+	OLEDFontWidth_3 = 3,
+	OLEDFontWidth_5 = 5, 
+	OLEDFontWidth_7 = 7, 
+	OLEDFontWidth_4 = 4, 
+	OLEDFontWidth_8 = 8,
+	OLEDFontWidth_16 = 16
+}OLEDFontWidth_e; // width of the font in bytes cols.
 
 typedef enum 
 {
-	FONT_O_EXTEND = 0x00, //   extends ASCII
-	FONT_O_SP = 0x20,  // Starts at Space
-	FONT_O_NUM = 0x30,  // Starts at number '0'
-}OLED_Font_offset_e; // font offset in the ASCII table
+	OLEDFontOffset_Extend = 0x00, //   extends ASCII
+	OLEDFontOffset_Space = 0x20,  // Starts at Space
+	OLEDFontOffset_Number = 0x30,  // Starts at number '0'
+}OLEDFontOffset_e; // font offset in the ASCII table
 
 typedef enum 
 {
-	FONT_H_8 = 8, FONT_H_16 = 16, FONT_H_32 = 32
-}OLED_Font_height_e; // width of the font in bits
+	OLEDFontHeight_8 = 8, 
+	OLEDFontHeight_16 = 16, 
+	OLEDFontHeight_32 = 32
+}OLEDFontHeight_e; // height of the font in bits
 
 class ERMCH1115_graphics : public Print{
 
@@ -75,8 +83,6 @@ class ERMCH1115_graphics : public Print{
 		int16_t radius, uint8_t color),
 	fillRoundRect(int16_t x0, int16_t y0, int16_t w, int16_t h,
 		int16_t radius, uint8_t color),
-	drawChar(int16_t x, int16_t y, unsigned char c, uint8_t color,
-		uint8_t bg, uint8_t size),
 	setCursor(int16_t x, int16_t y),
 	setTextColor(uint8_t c),
 	setTextColor(uint8_t c, uint8_t bg),
@@ -89,17 +95,19 @@ class ERMCH1115_graphics : public Print{
 
 	virtual size_t write(uint8_t);
 
-	void setFontNum(OLED_FONT_TYPE_e FontNumber);
+	void setFontNum(OLEDFontType_e FontNumber);
 	void drawCharNumFont(uint8_t x, uint8_t y, uint8_t c, uint8_t color , uint8_t bg);
 	void drawTextNumFont(uint8_t x, uint8_t y, char *pText, uint8_t color, uint8_t bg);
+	void drawChar(int16_t x, int16_t y, unsigned char c, uint8_t color, uint8_t bg, uint8_t size);
+	void drawText(uint8_t x, uint8_t y, char *pText, uint8_t color, uint8_t bg, uint8_t size);
+	
 	int16_t height(void) const;
 	int16_t width(void) const;
-	OLED_FONT_TYPE_e FontNum;
 
  protected:
 	const int16_t WIDTH;
 	const int16_t HEIGHT;
-	int16_t _width; 
+	int16_t  _width; 
 	int16_t  _height;
 	int16_t cursor_x; 
 	int16_t cursor_y;
@@ -111,10 +119,10 @@ class ERMCH1115_graphics : public Print{
 	bool wrap; // If set, 'wrap' text at right edge of display
 	bool drawBitmapAddr; // True = vertical , false = horizontal
 	
-	uint8_t _FontNumber = CH1115Font_Default;;
-	uint8_t _CurrentFontWidth = FONT_W_5;
-	uint8_t _CurrentFontoffset = FONT_O_EXTEND ;
-	uint8_t _CurrentFontheight = FONT_H_8;
+	uint8_t _FontNumber = OLEDFontType_Default;
+	uint8_t _CurrentFontWidth = OLEDFontWidth_5;
+	uint8_t _CurrentFontoffset = OLEDFontOffset_Extend;
+	uint8_t _CurrentFontheight = OLEDFontHeight_8;
 };
 
 #endif