+++ src/ER_OLEDM1_CH1115.cpp
@@ -7,7 +7,7 @@
 */
 
 
-#include "ER_OLEDM1_CH1115.h"
+#include "ER_OLEDM1_CH1115.hpp"
 #include <stdbool.h>
 
 // Class Constructors 2 off:
@@ -42,13 +42,17 @@ ERMCH1115  :: ERMCH1115(int16_t oledwidth, int16_t oledheight , int8_t rst, int8
 }
 
 
-
 // Desc: begin Method initialise OLED
 // Sets pinmodes and SPI setup
 // Param1: OLEDcontrast default = 0x80 , range 0x00 to 0xFE
-void ERMCH1115::OLEDbegin (uint8_t OLEDcontrast)
+// Param2: OLED_spi_divider default = 64 ,see bcm2835SPIClockDivider enum , bcm2835
+// Param3: SPICE_Pin default = 0 , which SPI_CE pin to use , 0 or 1
+void ERMCH1115::OLEDbegin (uint8_t OLEDcontrast, uint32_t OLED_spi_divider, uint8_t SPICE_Pin)
 {
 	_OLEDcontrast  = OLEDcontrast ;
+	_OLED_SPICLK_DIVIDER  = OLED_spi_divider;
+	_OLED_SPICE_PIN = SPICE_Pin;
+	
 	bcm2835_gpio_fsel(_OLED_RST, BCM2835_GPIO_FSEL_OUTP);
 	bcm2835_gpio_fsel(_OLED_CD, BCM2835_GPIO_FSEL_OUTP);
 	if(GetCommMode() == 3)
@@ -70,9 +74,21 @@ void ERMCH1115::OLEDSPIon(void)
 	bcm2835_spi_begin();
 	bcm2835_spi_setBitOrder(BCM2835_SPI_BIT_ORDER_MSBFIRST);
 	bcm2835_spi_setDataMode(BCM2835_SPI_MODE0);
-	bcm2835_spi_setClockDivider(BCM2835_SPI_CLOCK_DIVIDER_64);
-	bcm2835_spi_chipSelect(BCM2835_SPI_CS0);
-	bcm2835_spi_setChipSelectPolarity(BCM2835_SPI_CS0, LOW);
+	
+	if (_OLED_SPICLK_DIVIDER > 0)
+		bcm2835_spi_setClockDivider(_OLED_SPICLK_DIVIDER);
+	else // default BCM2835_SPI_CLOCK_DIVIDER_64 3.90MHz Rpi2, 6.250MHz RPI3
+		bcm2835_spi_setClockDivider(BCM2835_SPI_CLOCK_DIVIDER_64); 
+	
+	if (_OLED_SPICE_PIN == 0)
+	{
+		bcm2835_spi_chipSelect(BCM2835_SPI_CS0);
+		bcm2835_spi_setChipSelectPolarity(BCM2835_SPI_CS0, LOW);
+	}else if (_OLED_SPICE_PIN  == 1)
+	{
+		bcm2835_spi_chipSelect(BCM2835_SPI_CS1);
+		bcm2835_spi_setChipSelectPolarity(BCM2835_SPI_CS1, LOW);
+	}
 }
 
 // Desc: stop  Spi
@@ -466,15 +482,17 @@ void ERMCH1115::send_data(uint8_t byte)
 //Desc: updates the buffer i.e. writes it to the screen
 void ERMCH1115::OLEDupdate()
 {
-	OLEDBuffer( this->ActiveBuffer->xoffset, this->ActiveBuffer->yoffset, this->ActiveBuffer->width, this->ActiveBuffer->height, (uint8_t*) this->ActiveBuffer->screenbitmap);
-		return;
+	uint8_t x = 0; 
+	uint8_t y = 0; 
+	uint8_t w = this->_OLED_WIDTH; 
+	uint8_t h = this->_OLED_HEIGHT;
+	OLEDBuffer( x,  y,  w,  h, (uint8_t*) this->OLEDbuffer);
 }
 
 //Desc: clears the buffer i.e. does NOT write to the screen
 void ERMCH1115::OLEDclearBuffer()
 {
-	 memset( this->ActiveBuffer->screenbitmap, 0x00, (this->ActiveBuffer->width * (this->ActiveBuffer->height/ 8))  );
-	 return;
+	memset( this->OLEDbuffer, 0x00, (this->_OLED_WIDTH * (this->_OLED_HEIGHT /8))  ); 
 }
 
 //Desc: Draw a bitmap to the screen
@@ -525,33 +543,17 @@ void ERMCH1115::OLEDBuffer(int16_t x, int16_t y, uint8_t w, uint8_t h, uint8_t* 
 void ERMCH1115::drawPixel(int16_t x, int16_t y, uint8_t colour)
 {
 
-	if ((x < 0) || (x >= this->ActiveBuffer->width) || (y < 0) || (y >= this->ActiveBuffer->height)) {
+  if ((x < 0) || (x >= this->_OLED_WIDTH) || (y < 0) || (y >= this->_OLED_HEIGHT)) {
 	return;
-	}
-		uint16_t offset = (this->ActiveBuffer->width * (y/8)) + x;
-		switch (colour)
-		{
-		case FOREGROUND: this->ActiveBuffer->screenbitmap[offset] |= (1 << (y & 7)); break;
-		case BACKGROUND: this->ActiveBuffer->screenbitmap[offset] &= ~(1 << (y & 7)); break;
-		case INVERSE: this->ActiveBuffer->screenbitmap[offset] ^= (1 << (y & 7)); break;
-		}
-	return;
+  }
+	  uint16_t tc = (_OLED_WIDTH * (y /8)) + x; 
+	  switch (colour)
+	  {
+		case FOREGROUND:  this->OLEDbuffer[tc] |= (1 << (y & 7)); break;
+		case BACKGROUND:  this->OLEDbuffer[tc] &= ~(1 << (y & 7)); break;
+		case INVERSE: this->OLEDbuffer[tc] ^= (1 << (y & 7)); break;
+	  }
+
 }
 
-// Func Desc: init the Multibuffer struct
-// Param 1 Pointer to a struct
-// Param 2 Pointer to buffer array data(arrays decay to  pointers)
-// Param 3. width of buffer
-// Param 4. height of buffer
-// Param 5. x offset of buffer
-// Param 6. y offset of buffer
-void ERMCH1115::OLEDinitBufferStruct(MultiBuffer* mystruct, uint8_t* mybuffer, uint8_t w,  uint8_t h, int16_t  x, int16_t y)
-{
-   mystruct->screenbitmap = mybuffer; // point it to the buffer
-   mystruct->width = w ;
-   mystruct->height = h;
-   mystruct->xoffset = x;
-   mystruct->yoffset = y; 
-}
-
-//***********************************************
+// EOF